feat: fincube-5 Exchange user CRUD completed

+++ backend/user-management-service/src/exchange_user/exchange_user.controller.ts
@@ -1,34 +1,54 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Put, Param, Delete, NotFoundException } from '@nestjs/common';
 import { ExchangeUserService } from './exchange_user.service';
-import { CreateExchangeUserDto } from './dto/create-exchange_user.dto';
-import { UpdateExchangeUserDto } from './dto/update-exchange_user.dto';
+import { ExchangeUserEntity } from './entities/exchange_user.entity';
 
 @Controller('exchange-user')
 export class ExchangeUserController {
-  constructor(private readonly exchangeUserService: ExchangeUserService) {}
+  constructor(private readonly exchangeUserService: ExchangeUserService) { }
 
   @Post()
-  create(@Body() createExchangeUserDto: CreateExchangeUserDto) {
-    return this.exchangeUserService.create(createExchangeUserDto);
+  async create(@Body() exchange_user_entity: ExchangeUserEntity): Promise<ExchangeUserEntity> {
+    return this.exchangeUserService.create(exchange_user_entity);
+  }
+
+  @Post()
+  async findByEmail(@Body() email: string): Promise<ExchangeUserEntity> {
+    const exchange_user = await this.exchangeUserService.findByEmail(email);
+    if (!exchange_user) {
+      throw new NotFoundException('exchange_user not found');
+    } else {
+      return exchange_user;
+    }
   }
 
   @Get()
-  findAll() {
+  async findAll(): Promise<ExchangeUserEntity[]> {
     return this.exchangeUserService.findAll();
   }
 
+
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.exchangeUserService.findOne(+id);
+  async findOne(@Param('id') id: string): Promise<ExchangeUserEntity> {
+    const exchange_user = await this.exchangeUserService.findOne(+id);
+    if (!exchange_user) {
+      throw new NotFoundException(`Exchange user doesn't exist`);
+    } else {
+      return exchange_user;
+    }
   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateExchangeUserDto: UpdateExchangeUserDto) {
-    return this.exchangeUserService.update(+id, updateExchangeUserDto);
+  @Put(':id')
+  async update(@Param('id') id: string, @Body() exchange_user_entity: ExchangeUserEntity): Promise<ExchangeUserEntity> {
+    return this.exchangeUserService.update(+id, exchange_user_entity);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.exchangeUserService.remove(+id);
+  async remove(@Param('id') id: string): Promise<any> {
+    const exchange_user = await this.exchangeUserService.findOne(+id);
+    if (!exchange_user) {
+      throw new NotFoundException(`Exchange user doesn't exist`);
+    } else {
+      return this.exchangeUserService.remove(+id);
+    }
   }
 }
